+++ packages/docs/src/stories/Button.stories.tsx
@@ -22,33 +22,45 @@ export default {
         type: "inline-radio",
       },
     },
+    children: {
+      control: {
+        type: "text",
+      },
+    },
   },
 } as Meta<ButtonProps>;
 
-export const Primary: StoryObj<ButtonProps> = {};
+export const Primary: StoryObj<ButtonProps> = {
+  args: {
+    children: "Primary Button",
+  },
+};
 
 export const Disabled: StoryObj<ButtonProps> = {
   args: {
     disabled: true,
+    children: "Disabled Button",
   },
 };
 
 export const Small: StoryObj<ButtonProps> = {
   args: {
     size: "sm",
+    children: "Small Button",
   },
 };
 
 export const Secondary: StoryObj<ButtonProps> = {
   args: {
     variant: "secondary",
-    label: "New",
+    children: "Secondary Button",
   },
 };
 
 export const SecondaryDisabled: StoryObj<ButtonProps> = {
   args: {
     variant: "secondary",
+    children: "Secondary Button",
     disabled: true,
   },
 };
@@ -56,14 +68,14 @@ export const SecondaryDisabled: StoryObj<ButtonProps> = {
 export const Tertiary: StoryObj<ButtonProps> = {
   args: {
     variant: "tertiary",
-    label: "Cancel",
+    children: "Tertiary Button",
   },
 };
 
 export const TertiaryDisabled: StoryObj<ButtonProps> = {
   args: {
     variant: "tertiary",
-    label: "Cancel",
+    children: "Tertiary Button",
     disabled: true,
   },
 };